Added `index == -1` check to `encode:`
Added a `index == -1` check to `encode:`'s usage of 
`integerFromBinaryString:`.

The `integerFromBinaryString:` method can — in an error state — return 
`-1`, which would access `kEncodeChars[-1]`: invalid memory, observed 
to result in garbage characters (such as U+0005 ENQUIRY).

The fix checks for `-1` and will `@throw` a 
`NSInternalInconsistencyException`.  A throw may seem harsh but this 
should never happen, and if it does it's likely that either 
`integerFromBinaryString:`, the `NSRegularExpression` usage, or 
`binaryStringFromNumber:` needs to be fixed.  This resolves 
base32/issues#1